Remove org.eclipse.xtext project from classpath
From what I understand, having a reference to the org.eclipse.xtext
project this in the classpath only makes sense if you are also
developing xtext and have this project in your workspace.  In the common
case, however, we are going to use the xtext from the Eclipse
installation or that Gradle downloaded.